Correct problem with closured variable
Closured variable is evaluated at the time of execution. As the go routine
begine but wait for the Go scheduler to allow them to get executed, the for
loop ends, so the value of i will be the last one. The same issue applies to
Go channels, so be careful.

There is a pretty funny and unexpected way to solve this problem and it
involves some idomatic Go.