Race condition problem 1 is signified by the Previous write message.
After examining the related code, it is easy to see that the actual
problem is that the anonymous function takes no parameters, which means
that the value of i that is used in the for loop cannot be deterministically
discerned. as it keeps changing due to the for loop, which is a write operation.