Race condition 2
The data race is related to a write operation, and it happens on a Go map
by at least two goroutines. As the two goroutines have the same name(main.main.func1()),
this is an indication that they are the same goroutine. Two goroutines trying
to modify the same variable is a data race condition.