woker pool of goroutines with buffered channels
Client structure is assigned a unique ID for each request that are going
to be processed, the Data structure is used to group the data of a
Client with the actual results generated by the program. Put simply, the
Client structure holds the input data of each request, whereas the Data
structure holds the results of a request.

The global variables clients and data buffered channels are used to get
new client requests and write the results, respectively. You can
increase the value of the size parameter if you want the program to run
faster.

The worker() function reads the clients channel in order to get new
request to serve. Once the processing is complete, the result is written
to the data channel. The delay that is introduced using the time.Sleep()
is not necesary, but it gives you a better sense of the way that the
generated output will be printed. Finally, remember to use a pointer for
the sync.WaitGroup parameter in the worker() function, otherwise, the
sync.WaitGroup variable is copied, which means that it will be useless.